KolourPaint is yet another paint program.
It aims to be conceptually simple to use, while still providing a
reasonable level of functionality for the average user who just wants
to crop screenshots, draw diagrams and perform simple image editing
tasks.

This is not KIMP or Krita.  This is meant to be simple, but not so
simple that it lacks Undo/Redo.

It is intended to be an alternative to KPaint.

Currently (semi-) working features include:

- Selections
- Color Picker
- Pen
- Line
- Brush
- Eraser
- Spraycan
- Color Washer
- Rectangle, Rounded Rectangle
- Ellipse
- Fill
- Resize/Scale
- Flip
- Rotate
- Skew
- Convert to Greyscale
- Invert Colors

- Undo/Redo support (finally!!!)
- Zoom/Grid support
- Screenshot/Edit/Save/Set as wallpaper workflow optimisation
- Support for all file formats provided by KImageIO

The code is currently quite broken (well, maybe not as much as KPaint, but...)
but some people have requested code so here it is.  It still needs months of
work.

Most of the dialogs don't work, I know.

Oh, and don't complain about the ./configure script unless you're willing to fix it :)


svn path=/trunk/kdegraphics/kolourpaint/; revision=257753