Add icon for Polygon Tool.
It started out as a real polygon but for some reason, it turned out to look a
bit like mainland Australia.  So I decided to have a little "fun" and added
Tasmania to complete the picture :)

Ah, KolourPaint is turning out to be good for editing icons too :)
No transparency yet though and I need to actually _implement_ the Polygon Tool...  I think I'll do that now.

svn path=/trunk/kdegraphics/kolourpaint/; revision=258154