Fix bug where KolourPaint would refuse to draw anything because it thought that the mouse was not on top of a kpView:
QWidget::hasMouse() is _hopelessly_ unreliable after any kind of drag event.
So re-introduce that kpViewManager::setViewUnderCursor() hack and depend on
the kpView's to tell us what kpView the mouse is on top of.

It's not 100% reliable (switch to the Brush Tool and then do a dummy drag
of a tool onto the view and the brush pixmap cursor won't appear until you
move the mouse) but at least it will work whenever it's important (mousePressEvents).

svn path=/trunk/kdegraphics/kolourpaint/; revision=263619