The thumbnail makes a comeback!
Differences from the old thumbnail found in 0.2 (which eventually got
disabled because it was too buggy):

1. Caption shows zoom percentage
2. Maintains aspect ratio of image
3. Borders no longer consist of random garbage
4. Maximum zoom level is 100%:
   - for small images (e.g. icons), this is desirable as the mainView
     is probably set to a high zoom for pixel-level editing, while
     the thumbnail (@ 100%) displays what you will actually get
   - for large images, it would defeat the purpose of the mainView if
     the thumbnail provided >100% zoom; you are probably more
     interested in a true "thumbnail"
5. Added Close button
6. No longer docks at the most annoying times - in fact, docking has
   been completely disabled :)
7. Remembers its position

Now KolourPaint is not just a paint program but a full Icon Editor :)

Don't re-enable the Zoom actions when there isn't a document.
Work around a KSelectAction::setItems() bug in the process.

Rename kpMainWindow::slotImageMenuUpdateName() to
kpMainWindow::slotImageMenuUpdateDueToSelection() as it was also enabling
Image/Crop.

svn path=/trunk/kdegraphics/kolourpaint/; revision=279357