From the why-do-i-have-to-clone-everything-department:
kpSqueezedTextLabel is a source compatible (public methods) clone of KSqueezedTextLabel
with one major difference - it squeezes the tail instead of the middle of the string.

For URLs, squeezing in the middle makes sense but for (ignorable) status messages, "Dr...ri"
makes absolutely no sense in the statusbar.  But "Draw..." does (and even if it doesn't
it's at least less confusing).

Now I'm sure I've seen the squeezing-at-the-end behaviour in KDE apps before.
Is there a class that already does what I've just implemented?

svn path=/trunk/kdegraphics/kolourpaint/; revision=289660