kpDocumentSaveOptionsWidget = extension for filedialog for configuring save options (quality, colour depth) with an optional, realtime kpDocumentSaveOptionsPreviewDialog.
The great thing about the preview is that it gives you the actual size that
the file will be and what it actually looks like.  No more guesswork involved
in determining how much JPEG quality you want.  How does it work?  It saves the
image to memory and then reads it back.

This requires some changes to kpDocument that haven't been committed yet.

svn path=/trunk/kdegraphics/kolourpaint/; revision=329193