Avoid mentioning ver num to reduce maintenance hassle.
Explain why you should not and how to compile from source properly.

When I first used Linux, the most useless documentation I read was
the generic GNU INSTALL (which is retained at the bottom of this file):

./configure && make && make install

did not work for me because the generic instructions fail to explain
that ./configure fails if you are missing libraries and that make
fails if the build system is buggy.


svn path=/branches/kolourpaint/1.2_kde3/kdegraphics/kolourpaint/; revision=437497