Make renderer work on opaque images without 100% CPU usage:
- Remove Qt3 copyBlt() & bitBlt() calls which are badly emulated by Qt4 (either incorrect behaviour or very slow due to QPixmap -> QImage -> QPixmap translation)
- QPainter::drawRect() -> fillRect() to not be off by one
- Remove unused kpPixmapFX::getNonNullMaskAt(),setMaskAt()[2x],ensureOpaqueAt().  At last check, one of these was wrong anyway (untested code).

It _might_ be possible to use KolourPaint4 now depending on ones' interpretation of "working".

CCMAIL: Mike Gashler <gashlerm@yahoo.com>


Unfortunately, this patch is wrong (but better than the current state):

- Transparent image handling is ridiculously slow (QPixmap::mask() no longer returns a pointer in Qt4, it recalculates the mask and should there be an alpha channel (caused by simply using QPainter composition modes even if the alpha values will be either fully transparent or fully opaque), performs a QPixmap -> QImage -> QPixmap).  To get around this, use QPixmap::hasAlpha() to query for mask and don't ever try to get the contents of a mask.

- Preview dialog shows uninitialised pixels where the image should be transparent.  Don't know why.

- QPainter compositionModes do not work on QPixmap's when the X
server does not support XRENDER (XFree 3.3.6, VNC, Solaris/MacOS).
So we could simply avoid using compositionModes and do the old
scheme of drawing to the pixmap and then the mask (inconsistently enough, with XRENDER, it automatically draws to the mask).  Accessing
the mask is cheap if we avoid turning it into an alpha channel
(with composition modes).  However, in Trolltech's wisdom, they
removed raster ops so some mask operations will have to be emulated
SLOOOOOOOOOOOOWLY in software.  That's not too bad as:
1. The only place that really needs it is cutting out non-rect selections
2. Other places can simply avoid using mask operations.

But initial results without XRENDER suggest that drawing a line to a pixmap and then to the mask gives non-overlapping lines i.e. KolourPaint will not be able to run without XRENDER.  I will investigate further after my exams in a few weeks.

- Rectangles, rounded rectangles, ellipses and selections are still be off-by-one (and no, using QRect::adjusted (0, 0, -1, -1) does not work for a bounding rectangle 1x1 as it would become 0x0 - therefore it is impossible to use a QPainter::drawRect() to draw a 1x1 rectangle if a pen != Qt::NoPen!)

These problems are on top of the fact that all of KolourPaint is flawed since it's based on QPixmap.  I've had enough.  When I finish my investigations and port to Qt4, I will wrap all Qt painting functions acting on the document.  So next time, rectangles are made a further pixel wider and higher, or raster ops are dropped, or QPainter mask behaviour changes or subtlely, or QPainter acts differently with and without XRENDER (even without composition modes), I can simply change my code in a single place.  Also this will enable me to finally dump the insufficient QPixmap/QPainter backend.



svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=545207