1. Drag out a selection S 2. Invert / Clear / Convert To Greyscale etc. 3. Undo 4. Drag out another selection T 5. Redo 6. Undo
Should the current selection now be S or T?  The current behaviour is S
with the justification that Undo/Redo should only place you in states
you were in in the original undo history.

But T is handy because you want to avoid losing selection borders and
is also arguably correct.

I talked to Thurston and he concurs that T is crazy.  So I'm just
adding this patch to SVN in case we later change our minds.

svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=558455