Move duplicated add--create-selection-border-command--to-command-history logic into kpCommandHistory.
This moves kpToolSelectionCreateCommand::nextUndoCommandIsCreateBorder() into
kpCommandHistory as a file-local method.  This cleans up the design a bit since
commands should not be aware of the command history (but the other way around
is allowed).  The method also uses Q_ASSERT instead of "if" now.

kpCommandHistoryBase::{next,setNext}UndoCommand() are now only used by the subclass
kpCommandHistory, which is much cleaner, as those methods are hacks anyway.
But we don't change it from public to protected, since kpCommandHistoryBase
is supposed to be reusable outside of KolourPaint, without deriving a subclass.

svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=708923