Focus KNumInput's on creation instead of the OK button for the Skew and Rotate dialogs.  This allows the user to edit values without an extra mouse click.
For the Resize / Scale dialog, focus the "Width:" KNumInput instead of the Operation Group Box because:

Users probably want to start editing the dimensions straight away rather than the operation (the buttons inside the group box), which were obscure to change via keyboard anyway (did you know that you had to use the Left and Right arrow keys, rather than Tab?).  It is easy to select the operation afterwards with the mouse since the buttons are huge.

[BACKPORT of trunk/ -r720346 2007-10-03 and -r722282 2007-10-07]

svn path=/branches/KDE/3.5/kdegraphics/kolourpaint/; revision=722283