Fix resizing grip handles causing X11 size overflow
Instead of sizing the handles along the full extent of the view,
we just size them small enough for the handle to be visible. The
mask is now actually redundant, but I do not feel like rewriting
it.

Tested with 1600% zoom on 12000x9000 image, as well as image
smaller than grip size.

CCBUG: 237936

svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=1169574