Fix wrong JPEG 2000 MIME type
Together with fixes in kdelibs it is now possible to save
JPEG 2000 images in KolourPaint with adjustable quality.
Note that for JPEG 2000, the quality parameter really means
a file size percentage, so you should start with values
around 15, which already results in good quality.

If you do not get the "Quality" widget in the save dialog,
delete the kolourpaintrc file, as KolourPaint caches the
relevant MIME information in its config file.

svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=1215205