Add setting for anti-aliased drawing; lots of code cleanup
The mentioned bug is about rounded rectangle and anti-aliased drawing.
This patch fixes both: rounded-rectangles are now drawn with the same
radius in x and y direction, which makes it look much better and
avoids the mentioned problem in the bug report.
An additional option in the settings menu allows now to define if
the tools draw with anti-aliasing (default) or not.
BUG: 270203
GUI: added a new option in the settings menu for anti-aliased drawing