Port kpDocument_Save away from deprecated KIO::NetAccess
Still use a temporary file here, because the file might be big.

Ideally we should change savePixmapToDevice to write directly to the
transfer job, but then we need to change everything to be more async and
it quickly gets more complicated, so just keep it as close to the
original code as possible.