Delete custom kpToolAction class
This commit has no user-visible effect.

This commit deletes the kpToolAction class which is a KolourPaint-
specific subclass of KToggleAction.

kpToolAction used to have the specialised functionality of creating
a custom tooltip that showed the keyboard shortcut but this
functionality was removed in the previous commit because it
became redundant thanks to a frameworks change.

Because of this there is no reason left to keep this subclass
around. Instead its base class KToggleAction is used everywhere.